requirements.txt
Flask == 1.1.2
kaleido == 0.2.1
pandas == 1.2.4
plotly == 4.14.3
python-whois == 0.7.3
pywebio == 1.2.3
requests == 2.25.1
scikit-learn == 0.24.1
seaborn == 0.11.1
torch == 1.8.0
xgboost == 1.4.2
joblib == 1.0.1
matplotlib == 3.3.4
argparse == 1.4.0
tld == 0.12.5
beautifulsoup4 == 4.9.3
regex == 2021.4.4
urllib3 == 1.26.4

